Format and lint files in this repository
We want to maintain a consistent style in this repository to make it
easier to collaborate in it. For this purpose, linters and formatters
have been set up. They can be run via pre-commit hooks locally and are
run on every pull request to ensure the default branch is healthy.