Document how to test cache invalidations on Fastly
We use Fastly to serve Rust releases and crates. Sometimes it is necessary to
test that the cache can be purged, for which a new how-to guide has been
written.

The guide has been added to what will hopefully turn into a more complete
"service catalog" that supports the infra-team in its work.